requirements.txt
PyOpenGL
PyOpenGL_accelerate
pygame
numpy
Pillow